requirements.txt
streamlit>=1.32.0
scikit-learn>=1.4.0
numpy>=1.26.0
pandas>=2.2.0
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.20.0
umap-learn>=0.5.6
Pillow>=10.0.0
